requirements.txt
annotated-types==0.7.0
anyio==4.9.0
assemblyai==0.40.2
certifi==2025.4.26
chardet==5.2.0
distro==1.9.0
faiss-cpu==1.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
openai==1.79.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.0
reportlab==4.4.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
websockets==15.0.1
flask>=2.3.0
flask-cors>=4.0.0
openai>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.26.0
python-dotenv>=1.0.0 
